refactor(contact): unify form schema and extract sanitizer
Field validation lived in two hand-synced copies (form-types.ts and the
API route). Extract a shared CONTACT_FORM_FIELDS base object plus a shared
emailsMatch refinement; the client schema and the route now both derive
from it, with the route extending it for the Turnstile token.

Move sanitizeInput() into src/lib/utils/sanitize.ts so it is unit-testable
without importing the route.
