+++ src/app/api/contact/route.ts
@@ -1,82 +1,27 @@
 import sgMail from '@sendgrid/mail';
 import { NextResponse } from 'next/server';
-import sanitizeHtml from 'sanitize-html';
-import xss from 'xss';
 import { z } from 'zod';
 
 import { API_ERROR_MESSAGES } from '@/constants/config/api-error-messages';
 import { EMAIL_CONFIG } from '@/constants/config/email-config';
 import { CLOUDFLARE_API } from '@/constants/config/external-api-config';
-import { CONTACT_FORM_CONSTRAINTS } from '@/constants/config/form-config';
 import { getServerEnv } from '@/env';
 import { formatEmailTimestamp } from '@/lib/utils/date-utils';
 import { renderContactFormTemplate, generateContactFormText } from '@/lib/utils/email-template';
+import { sanitizeInput } from '@/lib/utils/sanitize';
+import { CONTACT_FORM_FIELDS, emailsMatch, EMAILS_MATCH_ERROR } from '@/types/form-types';
 
 import type { NextRequest } from 'next/server';
 
-// Build schema without refine to allow extension
-const ContactFormSchema = z
-  .object({
-    name: z
-      .string()
-      .min(CONTACT_FORM_CONSTRAINTS.NAME.MIN_LENGTH, 'Name must be at least 2 characters.')
-      .max(CONTACT_FORM_CONSTRAINTS.NAME.MAX_LENGTH, 'Name must be less than 100 characters.'),
-    email: z
-      .email('Please enter a valid email address.')
-      .max(CONTACT_FORM_CONSTRAINTS.EMAIL.MAX_LENGTH, 'Email address is too long.'),
-    confirmEmail: z
-      .email('Please enter a valid email address.')
-      .max(CONTACT_FORM_CONSTRAINTS.EMAIL.MAX_LENGTH, 'Email address is too long.'),
-    subject: z
-      .string()
-      .min(CONTACT_FORM_CONSTRAINTS.SUBJECT.MIN_LENGTH, 'Subject must be at least 3 characters.')
-      .max(
-        CONTACT_FORM_CONSTRAINTS.SUBJECT.MAX_LENGTH,
-        'Subject must be less than 200 characters.',
-      ),
-    message: z
-      .string()
-      .min(CONTACT_FORM_CONSTRAINTS.MESSAGE.MIN_LENGTH, 'Message must be at least 10 characters.')
-      .max(
-        CONTACT_FORM_CONSTRAINTS.MESSAGE.MAX_LENGTH,
-        'Message must be less than 2000 characters.',
-      ),
-    token: z.string().min(1, 'Turnstile token is required.'),
-  })
-  .refine((data) => data.email === data.confirmEmail, {
-    message: 'Email addresses must match.',
-    path: ['confirmEmail'],
-  });
+// Server schema = the shared client fields plus the Turnstile token.
+// Field validation lives in `@/types/form-types` (single source of truth).
+const ContactFormSchema = CONTACT_FORM_FIELDS.extend({
+  token: z.string().min(1, 'Turnstile token is required.'),
+}).refine(emailsMatch, EMAILS_MATCH_ERROR);
 
 const TurnstileResponseSchema = z.object({
   success: z.boolean(),
 });
-
-/**
- * Sanitizes user input to prevent XSS attacks and HTML injection
- * @param input - The string to sanitize
- * @returns Sanitized string safe for use in emails and responses
- */
-function sanitizeInput(input: string): string {
-  // First pass: strip all HTML tags
-  const htmlStripped = sanitizeHtml(input, {
-    allowedTags: [], // No HTML tags allowed
-    allowedAttributes: {}, // No attributes allowed
-    disallowedTagsMode: 'discard', // Remove disallowed tags entirely
-    allowedSchemes: [], // No URL schemes allowed
-    allowedSchemesByTag: {},
-    allowedSchemesAppliedToAttributes: [],
-    allowProtocolRelative: false,
-    enforceHtmlBoundary: false,
-  });
-
-  // Second pass: additional XSS protection
-  return xss(htmlStripped, {
-    allowList: {}, // No tags allowed
-    stripIgnoreTag: true,
-    stripIgnoreTagBody: ['script'],
-  }).trim();
-}
 
 export async function POST(request: NextRequest): Promise<NextResponse> {
   try {
